Create simple_python_email_bot
This is the 1.0 version of this bot. I plan on adding more features like reading a list of emails to send to, and a timer to send them in a timely manner. I used Outlook to avoid going through the "create app password on gmail", if you choose gmail have in mind that you will have to change the smtp_server to smtp.gmail.com, but the port remains the same. Some portuguese words in the middle to facilitate the naming hehe 

Initially I had A LOT of trouble separating the subject and body but it turns out that just breaking lines using \n AND writing 'Subject: " before, did the job... 
I'm also aware that there are ways to make this code safer but that wasn't my biggest concern at first. Cheers